app/requirements.txt
Flask==2.2.5
Werkzeug==2.3.8
gunicorn==22.0.0
psycopg2-binary==2.9.5
python-dotenv==0.21.1
prometheus-flask-exporter==0.22.0
python-json-logger==2.0.7